+++ utils/db/__init__.py
@@ -1,2 +1,3 @@
 from .instance import create_collections
 from .afk import afk_add_user, afk_get_user, afk_remove_user
+from .marry import marry_add_user, marry_get_user, marry_remove_user
