+++ utils/db/instance.py
@@ -5,5 +5,6 @@ db = json("db.json")
 def create_collections():
     for collection in (
         "afk",
+        "marry",
     ):
         db.create_collection(collection)
